requirements.txt
python-statemachine
pytest
pyautogui
pypiwin32
pywin32
python-telegram-bot
scipy
pillow
